Simulation/container/requirements.txt
boto3
pandas
cassandra-driver